feat: compaction
